+++ extensions/image-flow/src/process-file.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+import { Process } from "./process";
+import { getFiles } from "./supports/image";
+
+export default function Index() {
+  return <Process workflowName={"file"} imageLoader={getFiles} />;
+}